pyproject.toml
[project]
name = "startup-village-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp==3.9.1",
    "certifi>=2025.10.5",
    "flask==3.0.0",
    "google-generativeai==0.3.2",
    "jinja2>=3.1.6",
    "pillow==10.1.0",
    "python-dotenv==1.0.0",
    "reportlab>=4.4.4",
    "weasyprint>=66.0",
    "werkzeug==3.0.1",
]
